requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
lml==0.1.0
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
texttable==1.6.4
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
